requirements.txt
groq
pymupdf
python-docx
python-dotenv
streamlit